Added AuthService and UserController
