+++ src/main/java/com/example/auth_api/repository/UserRepository.java
@@ -2,7 +2,6 @@ package com.example.auth_api.repository;
 
 import com.example.auth_api.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
