+++ src/main/java/com/example/auth_api/services/AuthService.java
@@ -1,0 +1,8 @@
+package com.example.auth_api.services;
+
+
+public interface AuthService {
+
+    boolean isLogInSuccess(String username, String password);
+
+}
